Add level to wild grass and surf

+++ Source/Core/Gen8/EncounterArea8.hpp
@@ -27,7 +27,7 @@ class EncounterArea8 : public EncounterArea
 public:
     EncounterArea8() = default;
     EncounterArea8(u16 location, Encounter type, const std::vector<Slot> &pokemon);
-    u8 calcLevel(u8 index, u16 prng) const;
+    u8 calcLevel(u8 index, u32 prng) const;
     u8 calcLevel(u8 index) const;
 };
